Time: 288 ms (95.03%), Space: 117 MB (89.70%) - LeetHub

+++ 295-find-median-from-data-stream/295-find-median-from-data-stream.cpp
@@ -28,7 +28,7 @@ public:
     double findMedian() {
         if( size % 2 != 0 )
             return minHeap.top();
-        return double(minHeap.top() + maxHeap.top())/2;        
+        return (minHeap.top() + maxHeap.top())/2.0;        
     }
 };
